sign-recognizer-backend-main/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1.1
djangorestframework==3.15.2
gunicorn==23.0.0
whitenoise==6.7.0
psycopg2-binary==2.9.9
sqlparse==0.5.1
requests==2.32.3
sentry-sdk==2.19.2
django-csp==3.8
urllib3==2.2.3
django_cloudinary_storage==0.3.0
opencv-python-headless==4.5.5.64
mediapipe==0.10.5  # Daha hafif bir versiyon
numpy==1.24.3  # Python 3.11 ile uyumlu versiyon
pandas==2.0.3  # Python 3.11 ile uyumlu versiyon
scikit-learn==1.3.2  # Modelin eğitildiği sürümü kullanın
joblib==1.2.0  # Python 3.11 ile uyumlu versiyon